+++ src/Router.tsx
@@ -1,3 +1,14 @@
+import { createBrowserRouter, RouterProvider } from 'react-router-dom';
+
+import { MainPage } from '@/pages';
+
 export default function Router() {
-  return;
+  const router = createBrowserRouter([
+    {
+      path: '/',
+      element: <MainPage />,
+    },
+  ]);
+
+  return <RouterProvider router={router} />;
 }
